packages/lambda/docx-mcp/pyproject.toml
[project]
name = "idp_v2.docx_mcp"
version = "1.0.0"
description = "DOCX MCP tools for text extraction, table extraction, and DOCX creation"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "boto3>=1.42.0",
    "python-docx>=1.1.0",
    "nanoid>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [".ruff_cache", ".svn", ".tox", ".venv", "dist"]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.pyright]
typeCheckingMode = "basic"
reportArgumentType = false
reportUnusedCallResult = false
reportExplicitAny = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnannotatedClassAttribute = false
reportMissingTypeStubs = false
reportCallInDefaultInitializer = false
reportInvalidTypeForm = false
reportAny = false
reportMissingParameterType = false
reportImplicitOverride = false
reportAttributeAccessIssue = false
